feat(cmd): create basic SysCommand class
The SysCommand class allows us 'shell out' and run commands on the native machine.

It uses forkpty() to create a slave process to run the given command in an emulated terminal (pty).
Epoll is used to monitor IN and HUP events from the slave file descriptor.

If the exit status is zero, the execute() function reports a success.